Run dusk in docker on travis ci (#49)
Trying to make Dusk work on Travis and Circle CI

This is still not working and I have decided to leave it for now, as the Dusk
tests all pass in my local (with Selenium), to concentrate on upgrading
Laravel. But since I have done some useful changes, I would like to keep
them and therefore merging them into master.