Refactor authenticaed routes in their own group
The /dashboard route is the only one, so far, that needs authentication.
It added the auth middleware in the controller, but since I will have
a lot more routes which will need an authenticated user it's easier to
add the middleware in the routes file rather than in each single
controller.