Fix the password reset flow
The standard Laravel email to reset a password send the URL with the
email address as a query parameter. I personally consider this a
security risk and therefore I have removed it. The user is still asked
to enter the email address they want to reset the password for which
has to match the one the reset link was requested for.