Seed the database
I use to rely to manually enter data in the site to progress to the
next pages and see how things look. This is not very scalable, as
I will need a lot of data (think about the fixtures for example).
I need a way to create a lot of data consistently and automatically.