Refactor the listeners so we can add more jobs
The listeners that create the roles when a season, competition, etc, are created will need to do more. So, rather that just create the role, they have been refactored to call a job to create the role. This way we will be able to add more jobs to create the permissions and attach the permissions to the roles.